refactor: generalize interpolatable templates and views
